0.26.0 - Created TreeNode interface.  Moved Translation and ListOfTranslation to new bast package
