+++ src/main/java/com/bashpile/engine/bast/TreeNode.java
@@ -1,0 +1,6 @@
+package com.bashpile.engine.bast;
+
+public interface TreeNode<T> {
+    T getData();
+    TreeNode<T> add(TreeNode<T> child);
+}
