+++ src/main/java/com/bashpile/engine/bast/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * For the Bashpile Abstract Syntax Tree
+ */
+package com.bashpile.engine.bast;
